Replace use-ok by simple use in a test file
The design of Perl 6 requires that use statements only have effect in a
lexical scope. Otherwise it's impossible to use different (versions of)
modules sharing a name in different parts of a program.

This means essentially that we need to explicitly use all modules we need
in a scope. And that the use-ok can only affect the BEGIN block itself,
not the rest of the test file.

See https://gist.github.com/niner/70f7b46eefb7e22af78d896bea11efeb for
further information.